requirements.txt
torch==2.4.0
transformers==4.57.1
accelerate>=0.30.0
einops>=0.7.0
huggingface-hub>=0.23.0
Pillow>=10.2.0
qwen-vl-utils>=0.0.8
safetensors>=0.4.3
sentencepiece>=0.1.99
tiktoken>=0.7.0
numpy<2
opencv-python>=4.8.0
